Miki: What's up Neveloper? Neveloper: I am trying to transform agile into a humanish agileish approach which brings the team's well-being into the focus. In my opinion the best way to increase the team's productivity by making them happy. Nikolai: Buy them a table-soccer, everybody knows table-soccer is the only way to make developers happy. Miki: It is only one factor from many, but I do like to play table-soccer with my colleagues, but of course it is definitely not for everyone.
I love the humanish agileish method, we use that and it works perfectly.
Neveloper: Of course, it works. I came up with the idea.
Miki: Actually, it was me.
Neveloper: That might be true, but you had no balls to post it under your name and you gave life to me just to make sure they won't throw tomato on you.
Miki: Well. That is partially true. But most importantly I hired you as you are the expert in software development. Having talks with you gives life to new ideas, I really enjoy talking to you.
Minnie: Awww.